requirements.txt
Flask==2.3.2
nltk==3.8.1
numpy==1.25.2
openai==0.27.6
pandas==1.5.2
pdfminer.six==20221105
regex==2023.5.5
